Move postgres schema features to SchemaDialect
